\name{dataset.all.species}
\docType{data}
\alias{dataset.all.species}
\title{Angiosperm Dataset}
\description{
This data set gives an example of angiosperm species and their occurence in southern america.
}
\usage{dataset.all.species}
\format{A dataframe containing the longitude and lattitude of the occurence location as 'long' and 'lat' and the species ID as 'speciesID'.}
\source{Angiosperm Database}
\references{
Morawetz, W. and Raedig, C. (2007). Angiosperm biodiversity, endemism and conservation in the Neotropics. Taxon 56:1245-1254.
}
\keyword{datasets}
