\name{dataset.height}
\docType{data}
\alias{dataset.height}
\title{Height-Information Dataset}
\description{
This data set containes the height of 1 degree grid cells of southern america.
}
\usage{dataset.height}
\format{A dataframe containing the longitude and lattitude of the grid cell location as 'long' and 'lat' and the height of the cell as 'height'.}
\keyword{datasets}
