\name{dataset.landwater}
\docType{data}
\alias{dataset.landwater}
\title{Land-Water-Information Dataset}
\description{
This data set containes the percentage of land/water of 1 degree grid cells of southern america.
}
\usage{dataset.landwater}
\format{A dataframe containing the longitude and lattitude of the grid cell location as 'long' and 'lat' and the percentage of land/water as landsum.}
\keyword{datasets}
